# Filter scheme expressions before testing VAT number