| Q             | A
| ------------- | ---
| Bug fix?      | no
| New feature?  | no
| BC break      | no
| Tests passed? | yes
| Fixed tickets | none
| License       | MIT

Small change in performance. It only test the eligible expressions for the given country, not the complete list.