@antalaron 2018-07-19
Thank you very much!

I've made some performance checks:

Before:
```
$ time vendor/bin/phpunit >/dev/null

real	0m1,696s
user	0m1,582s
sys	0m0,095s
```

After
```
$ time vendor/bin/phpunit >/dev/null

real	0m1,620s
user	0m1,510s
sys	0m0,092s
```

It really improves the code.

I have some request.